Perform /f top async + more optimization
We should always use our new economy cache for the summation of a
factions money since we save a lot of IO.

We run /f top async because it does not change the state of any
components, it merely reads what exists, calculates the top, and tells
the user. All of which is thread safe.